+++ src/toisto/model/language/grammar.py
@@ -8,7 +8,16 @@ GrammaticalNumber = Literal["infinitive", "singular", "plural"]
 GrammaticalPerson = Literal["first person", "second person", "third person"]
 DegreeOfComparison = Literal["positive degree", "comparitive degree", "superlative degree"]
 Tense = Literal["present tense", "past tense"]
-GrammaticalCategory = Literal[GrammaticalGender, GrammaticalNumber, GrammaticalPerson, DegreeOfComparison, Tense]
+SentenceType = Literal["declarative", "interrogative"]
+GrammaticalCategory = Literal[
+    GrammaticalGender,
+    GrammaticalNumber,
+    GrammaticalPerson,
+    DegreeOfComparison,
+    Tense,
+    SentenceType]
 
-# A plural concept automatically uses the singular form and past tense automatically uses the present tense
-AUTO_USES = {"plural": "singular", "past tense": "present tense"}
+# Mapping of grammatical categories to grammatical categories they automatically use. For example, a plural concept
+# automatically uses the singular form. This forces Toisto to present quizzes of the used category before presenting
+# quizzes for the using category.
+AUTO_USES = {"plural": "singular", "past tense": "present tense", "interrogative": "declarative"}